refact(isReady): moved isReady to it's own file
